+++ notecli.gemspec
@@ -2,7 +2,7 @@
 
 Gem::Specification.new do |spec|
   spec.name          = "notecli"
-  spec.version       = "0.0.1"
+  spec.version       = "0.0.2"
   spec.authors       = ["zpallin"]
   spec.email         = ["zpallin@gmail.com"]
